Prevent illegal N+1 multi-card drop on empty stack in Forty Thieves
PROBLEM:  In Forty Thieves, if there are N free spaces (empty stacks)
          available then the game permits pickup of N+1 cards.  It is
          valid to drop N+1 cards on a non-empty stack.  However,
          dropping N+1 cars on a free space (empty stack) is not
          valid, and was permitted.

          It is valid to drop a maximum of N cards on a free space
          (empty stack).

SOLUTION: One option to address the illegal multi-card drop would be
          to limit pickup to N cards, or even 1 card.  However this
          restricts flexibility in game play.  Instead limit valid
          drop actions on free spaces (empty stacks) to N cards.

          This solution uses a modified patch from issue 29, see
          references.

REFERENCES: This problem was reported as issue 29 in the
            solitaire-for-android issue tracker.

Issue 29:  illegal multi-card moves in Forty Thieves
https://code.google.com/p/solitaire-for-android/issues/detail?id=29

Following are the details from the issue report:

   What steps will reproduce the problem?

   1. Play until a space has been achieved and there is a 2-or-more
      card sequence in suit (e.g. 4s-5s, or 3h-4h-5h, etc.)
   2. move two cards from sequence into space
   3. Notice the two cards successfully moved to space.

   What is the expected output? What do you see instead?

   The expectation is that this move is illegal.  When you have a
   space, any single card can be moved there.  This can be used to
   move two cards to another target stack, such as 4s-5s moved to 6s,
   by moving 4s->space, 5s- >6s, 4s->5s.  In general you can move n+1
   cards when n spaces are present.  However, if the target is a
   space, then only n cards can be moved.

An emptystackpatch.patch file was provided by sirpe...@gmail.com.
Unfortunately I was unable to view the full email address and hence
could not provide full credit for the original author of this patch.

In my testing the emptystackpatch.patch prevented the drop of any
cards on a free space (empty stack).  However, with a minor logic
change this patch addressed the issue.