Show restart unavailable message if game won and restart selected
PROBLEM:  Game crashes or hangs if restart selected when game won and
          animated card replay in action.

          The steps to reproduce the problem are:

          1. Play and win a game.
          2. While animated card replay in action, choose Menu -> Restart
          3. Game freezes.

SOLUTION: Add a flag to indicate if game won or not.  If game won,
          show message "unable to restart" game.  Otherwise if game
          not won (game in progress), perform restart action.

REFERENCES: This problem was reported as issue 45 and issue 50 in the
            solitaire-for-android issue tracker.

Issue 45:  Restarting when winning a game
https://code.google.com/p/solitaire-for-android/issues/detail?id=45

A patch was included in issue 45 by maca...@gmail.com.  Unfortunately
I was unable to view the full email address and provide additional
credit.  The issue 45 patch added a flag to indicate game won status,
and then would disable the restart menu if game won.

The game won status portion of the patch was used; however, instead of
disabling the restart menu I opted to display an "unable to restart"
message if the game had been won.