Restore automatic save of game statistics and settings
Version 2.6.1 broke the automatic saving of game statistics and
settings.  This meant that an upgrade to a newer version, such as 2.7,
started as if a fresh install.  None of the previous game statistics
or settings survived the upgrade.

The cause was a side-effect of the following commit:

    Restore ability to undo moves after application exit
    7f68d76fc16a71e094d0443fa28af0551896cc59

Fix by ensuring that SaveGame() is invoked if the application reaches
the onStop() state.

Note that the issue is also present in version 2.7.

Ticket #12 Statistics and settings lost after version update
https://sourceforge.net/p/solitairecg/tickets/12/