Fix freecell logic for single step move of a valid card sequence
A freecell powermove is simply a shortcut move that lets you move a
valid sequence of cards in a single move.  The same move can be made
by moving cards one at at time to achieve the same result.

The number of cards that can be moved in a powermove in freecell is
calculated as follows:

(1 + number of empty freecells) * 2 ^ (number of empty tableaus)

Following is a chart which demonstrates this relationship:

                       Card
Empty      Empty       Sequence
Tableaus   Freecells   Length
--------   ---------   --------
    0          0          1
    0          1          2
    0          2          3
    0          3          4
    0          4          5
    1          0          2
    1          1          4
    1          2          6
    1          3          8
    1          4         10
    2          0          4
    2          1          8
    2          2         12
    2          3         16
    2          4         20

This assumes the sequence is moved to a non-empty tableau.  If the
move is into an empty tableau then this empty tableau does not count
as an empty tableau for the formula above.