Fix forty thieves logic for single step move of a valid card sequence
A forty thieves powermove is simply a shortcut move that lets you
move a valid sequence of cards in a single move.  The same move can be
made by moving cards one at at time to achieve the same result.

The number of cards that can be moved in a powermove in forty thieves
is calculated as follows:

2 ^ (number of empty tableaus)

Following is a chart which demonstrates this relationship:

           Card
Empty      Sequence
Tableaus   Length
--------   --------
    0         1
    1         2
    2         4
    3         8
    4        16
    5        32

This assumes the sequence is moved to a non-empty tableau.  If the
move is into an empty tableau then this empty tableau does not count
as an empty tableau for the formula above.